pyproject.toml
[project]
name = "aurora-sun-v1"
version = "0.1.0"
description = "AI coaching for neurodivergent people"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}

dependencies = [
    # All dependencies pinned to exact versions for reproducibility.

    # Core
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "python-telegram-bot==21.10",
    "sqlalchemy==2.0.36",
    "alembic==1.14.1",
    "asyncpg==0.30.0",

    # AI/LLM
    "anthropic==0.43.0",
    "openai==1.59.6",
    "groq==0.13.1",

    # LangGraph & DSPy (pre-release: API may change between versions)
    "langgraph==0.2.62",        # pre-release: pin strictly
    "dspy-ai==2.5.43",          # pre-release: pin strictly
    "pydantic-ai==0.0.39",      # pre-release: pin strictly

    # Vector & Graph
    "neo4j==5.27.0",
    "qdrant-client==1.12.1",

    # Memory
    "letta-client==0.6.5",

    # Redis
    "redis==5.2.1",

    # Security
    "cryptography==42.0.8",
    "keyring==25.5.0",
    "PyJWT==2.9.0",

    # Monitoring
    "langfuse==2.58.2",
    "prometheus-client==0.21.1",

    # Utilities
    "pydantic==2.10.4",
    "python-dotenv==1.0.1",
    "httpx==0.28.1",

    # Logging
    "structlog==24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "ruff==0.8.4",
    "mypy==1.13.0",
    "bandit==1.8.2",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
